worker/requirements.txt
pypdf2
vertexai
pymongo
python-dotenv
